cli: `zucc load canvas': Initialize submission metadata
Also swap from yaml to json for the metadata because in theory, this
metadata doesn't need to be human-editable. It's just a convenient
format for saving metadata.

I'm thinking we also save grading information, possibly including
logging, to this file. Downside is that it makes parsing the file pretty
expensive if logs get big (and they can when valgrind gets out of
control.)